add auth protection in backend on modifing and deleting post

+++ backend/routes/posts.js
@@ -64,10 +64,15 @@ router.put(
       content: req.body.content,
       imagePath: imagePath,
     });
-    console.log(post.toObject());
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      res.status(200).json({ message: "Update successful!" });
-    });
+    // will update when both condition are matched
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
+      .then((result) => {
+        if (result.nModified > 0) {
+          res.status(200).json({ message: "Update successful!" });
+        } else {
+          res.status(401).json({ message: "Not authorized!" });
+        }
+      });
   }
 );
 
@@ -105,9 +110,13 @@ router.get("/:id", (req, res) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if (result.deletedCount > 0) {
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
   });
 });
